+++ src/main/java/com/lhy/ae2utility/jei/EncodePatternButtonController.java
@@ -13,7 +13,6 @@ import org.lwjgl.glfw.GLFW;
 import com.mojang.blaze3d.platform.InputConstants;
 import com.lhy.ae2utility.compat.JeictCompat;
 import com.lhy.ae2utility.compat.WcwtCompat;
-import com.lhy.ae2utility.compat.SomeUselessThingsCompat;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.world.inventory.AbstractContainerMenu;
@@ -556,10 +555,6 @@ public class EncodePatternButtonController implements IIconButtonController {
                             true);
                 }
             }
-        }
-
-        if (SomeUselessThingsCompat.transfer(recipeLayout, true)) {
-            return true;
         }
 
         boolean shiftDown = Screen.hasShiftDown();